created transaction update component to modify/remove transaction
